Prevent overriding session store key in test environment
In the test environment if the session store is overriden it causes
the Capybara tests to misbehave; essentially clearing the session
between each request (resulting in the sign up flows not working as the
order depends on session data from previous steps). I'm not entirely
sure why yet; there may be a better way to fix the issue by configuring
Capybara.