Create models/views/wizard representing linear flow
Re-purposes the existing TTA sign up models and views; consolidating
them so the same steps are reused throughout (instead of having a
different model for the various branches, as the steps don't actually
change). The journey is then controlled by a single wizard, with the
steps determining if they should be shown or not depending on the data
in the store from previous steps.