Allow submission of steps with no attributes
Some of the steps, such as `ReviewAnswers`, do not send any data back on a form
submission. Also, some questions only have a radio button which, when
unselected, results in no data being sent to the server. To account for this we
check that the params contain the expected root key before calling `require`
(preventing an exception when they don't).