Expose reviewable_answers from individual steps
Returns the reviewable answers from each step; this is complicated by the fact
that some steps have multiple attributes that are aggregated for review (such
as `UkAddress#address`) and others have attributes that are 'hidden' and not
set explicitly by the user (such as `HaveADegree#degree_type_id`).

The partial for rendering a step then becomes simple - the headings for each
answer are pulled out of translations.