Move view logic out of models
Instead of having the models responsible for formatting dates the
`reviewable_answers` have been updated to return `Date` or `Time` instances
that are then formatted in a helper.

At the moment any `Time` object will be formatted as `HH:mm` even though a
`Time` object contains date information - if you want to output it as a date it
needs to be a `Date` object (so use the `to_time` and `to_date` methods on the
`Date/DateTime/Time` classes).