Prevent candidate from signing up multiple times
If a candidate has already signed up to the TTA service and successfully
matches back, the store will contain a `true` value for
`already_subscribed_to_teacher_training_adviser`. When this is the case, we
will show them a new step that informs them they have already signed up and
directs them to contact support if they think this incorrect.