Display review answers correctly when steps have been missed
It is currently possible for a user to miss a step; you can jump straight to
the review answers step if you have the URL (though if you try and complete the
submission it will redirect to the first invalid step).

If a candidate does this it causes an exception due to not expecting a
non-skipped step to have a `nil` value (which happens if they manually plug a
future step URL in). This commit updates the `reviewable_answers` to ensure an
exception will never be raised, instead displaying an empty answer in the
review answers step for any they have missed.