Update callback helper to use Time.zone
The `Time.zone` is now set per-request for the wizard steps; this commit
updates the helper to make use of this time zone instead of it being explicitly
passed in.

Also extracts the business logic to exclude the current day out of the view
helper and into the model concern.