Prevent API check when already authenticated
When the wizard is about to be completed we check for any invalid steps;
this was resulting in another API call to verify the verification code in the
`Authenticate` step. If the user takes too long completing the form their code
may not be valid by the time they reach they come to submit it, resulting in
them being redirected back to the `Authenticate` step.

This commit sets an `authenticated` flag in the store; if this is present
the `Authenticate` step will not try and re-verify their code.