Custom error for invalid date of birth
Previously when an invalid `date_of_birth` was set we were showing the
validation error associated with the `presence` validation (as we rescue and
`nil` the `date_of_birth` if it can't be parsed as a `Date`).

Instead, this commit will add a custom error message when we rescue from an
invalid date format and inform the user that they did not enter a valid date of
birth.

Clean up `DateOfBirth` model, specifically around attribute assignment in
`date_of_birth`.