Remove callback booking steps for overseas/equivalent degree
At the moment if a candidate has an equivalent degree and is located
overseas we prompt them to book a callback by asking for their telephone
number and time zone, then displaying a list of available slots to book.

TPUK do not currently have the ability to make outbound international phone
calls, so we are unable to offer scheduled callbacks for these candidates.
Instead, we will prompt the candidate to enter their overseas telephone
number.

The step has been kept as `overseas_time_zone`, which is a little
misleading but it should make reverting the PR later cleaner (we're
planning to revert this as soon as TPUK support outbound international
calls).