Remove redundant validation checks
These attributes are not set by the user; instead they are set by the model
internally. If these validation checks ever failed we would end up showing an
error message to the user that they could do nothing about, and we would not be
alerted to a bug in the application code.

Instead, by removing the validation checks it will enable the candidate to get
to the end of the submission and they will then receive an error when the form
attempts to submit to the API. We will also be alerted about this, which feels
like a better outcome/experience for both the candidate and ourselves.