Send CSP violations to Logit.io
- Add CspReportsController

Introduce a CspReportsController that accepts a CSP violation report (any JSON
payload with a csp-report key) and logs it as an error.

- Report CSP violations via Rails

Instead of the CSP violations going directly to Sentry, we're going to send
them instead to Rails where they get logged out and will end up in logit.io. We
can then setup a Grafana dashboard to track them.

Any requests with a malformed body are ignored, hopefully reducing the
exploitability a little.

- Whitelist CSP keys and cap entry length

To help prevent malicious usage of the endpoint we ensure the csp-report only
contains valid keys and truncate the length of all values (hopefully avoiding
someone maliciously filling up our logstash).

I thought about setting up the CSP report as a model to add further validation
(ensure the blocked-uri is a URL, for example) but it feels like more hassle
than it's worth.

- Rate limit CSP reporting

As the CSP reporting endpoint allows POST requests directly from clients that
end up in our logs, we want to restrict how frequently it can be called by a
client.

Restricts POST /csp_reports to 5rpm per IP address.