Rescue ApiError and render 429 error page
Whilst the app itself is rate limited the API also has rate limiting. In the
unlikely event that the client rate limits don't trigger and instead the API's
rate limiting is enforced we want to display the same 429 error page to the
user.

Raise `ApiError` up to the `ApplicationController` when the `code` is 429. The
`ApplicationController` will then rescue the error and render the error page.