Prevent error when skipping ahead to review answers step
If a user proceeds through the TTA sign up as a candidate with an equivalent
degree that is located in the UK, but then manually enters the URL for the
review answers page before completing the `UkCallback` step an error would be
thrown as the `phone_call_scheduled_at` is `nil` and we try and call `to_date`
on it.

Update to use the safe navigation operator on `phone_call_scheduled_at`.