Redirect to identity step on resend_verification bad request
If a candidate's session expires (so they sit on the authenticate step
overnight, for instance) then they hit the link to resend their verification
code the app attempts to make an invalid request to the API (email, first and
last name are nil). This currently results in a 500 error for the user.
Instead, we rescue this condition and redirect to the identity step so that
they can re-authenticate.