Improve user experience when SubjectLikeToTeach is not found
Change the "Can't see the subject you would like to teach?" prompt to be an
extra radio button on the "What subject would you like to teach if you return
to teaching?" question.

Add an additional step to handle when the user selects the new radio option
(similar to the `QualificationRequired` step).

Tidy up partial for rendering radio buttons.

Improve wording of error message to make sense with the new option value.

Extend the `ApiOptions` helper with support for including only a subset of the
types returned from the API.