Remove redundant attribute from ReturningTeacher step
This was previously needed to default the `preferred_education_phase_id` to
secondary when they are a returning teacher. This logic was moved to the API
and subsequently removed from the app by PR #413 but the attribute was left in.
This causes a bug when the user:

- Starts down the non-returner path
- Selects Primary as their education phase
- Backs out and goes down the returner path
- Completes the wizard

The `preferred_education_phase_id` is then submitted to the API as Primary
(from the `ReturningTeacher` step) when it shouldn't be sent at all, as we
default the value in the API now.