Prevent completing sign up on skipping non-proceedable step
Currently it is possible for a candidate to hit a 'dead end' and then manually
skip to the review answers screen (which they may have open in another tab) and
attempt to complete the sign up, which will result in a 500 error.

This is possible because we currently only check that all 'active'
(non-skipped) steps are valid, but we neglect to check if they are a 'dead end'
step themselves. An example of this would be:

- Returning candidate selects 'Other' as the subject they would like to teach, which is stored as `-1` in the `preferred_teaching_subject_id` attribute (and valid, otherwise they would not to the 'Get support' step).
- Candidate is presented with the 'Get support' dead-end step.
- Candidate manually skips ahead to the review answers screen, is prompted to fill in any other steps they have skipped and can complete the sign up.
- `-1` is submitted to the API for `preferred_teaching_subject_id` and an error is raised as this is not a valid GUID.

To prevent this, we introduce another check on completion to ensure the
candidate was able to proceed through all valid steps. If this is not the case
the candidate gets redirected to the dead-end step, explaining why they can't
proceed (at which point they can choose to go back and change their answer in
order to continue).