Simplify Authenticate step with access_token_used? helper
The `Wizard::Base` now contains a method to determine if/what kind of
authentication method has been used.

Instead of manually tracking this information in the `Authenticate` step we can
remove the logic and call out to the wizard.

Ensure `timed_one_time_password` dirty tracking is cleared on `ApiError` so
that if a user enters a bad code we only try and validate it with the API once
(unless they change the code and try again).