Add crm data support to store
We want to hide steps that are optional and have been
pre-filled with candidate information.

In order to determine if a step can be pre-filled from the
CRM we need to retain the CRM values in the store so that they
can later be queried independently of the data written by the
app.

The `fetch` method has also been updated to return values for
all keys passed in; this makes the logic to determine if all
values are present in a store simpler (otherwise we would also
have to check the `count` to ensure all values are present):

```
fetch(%[...], source: :crm).values.all(&:present?)
```