Add support for optional steps
An optional step will be skipped if (and only if) the step attributes were all
pre-filled from the CRM data.

Data will be send to the API for skipped steps going forward.

This is going to be useful because skipped step values will be `nil` for new
candidates and will contain the data already in the CRM for existing
candidates. Ultimately this will make it easier to update the API to support
sending `nil` values to the CRM (it avoids having to duplicate the logic for
'was this step shown to the user' in the API).