Update ReturningTeacher step to send type_id
In the past we haven't sent a specific attribute value in response to the
`ReturningTeacher` step (the previous `returning_to_teaching` attribute was
transient and only used to control the flow of the following wizard steps).

The API now accepts a `type_id` as part of the `TeacherTrainingAdviserSignUp`
request model, so we will populate this on the `ReturningTeacher` step.

The existing `returning_to_teaching` attribute has been removed but a method
that behaves the same takes its place, avoiding further refactoring (really it
should be `returning_to_teaching?` but this PR is big enough so we can pick up
that refactor later on).