Show telephone field on review answers step
The telephone field when filled in as part of the `OverseasTimezone` step was
not presented under the personal details section; its now correctly located.

The `UkCallback` step has a mixture of personal/non-personal details and was
only showing the non-personal (callback date/time) - it has been updated to
render the telephone as well and the answers have been shifted into the
personal details section.

Going forward it would make sense to split telephone/callback details into
their own section perhaps (instead of trying to split a single steps answers
across multiple sections when it contains a mixture of personal/non-personal
details).