Fix bug where NoDegree step is shown incorrectly
If a candidate initially starts their journey as a non-returner and then
selects no degree we will show them the exit step informing them that they have
to have a degree.

If they they go back and try to carry on as a returner they will see the same
exit step. Instead, we should be ignoring their selected degree option when on
the returner journey.