Bug fix callback scheduled for returner
If a candidate goes all the way through the equivalent degree/UK path and
schedules a phone call, then edits the answer for 'are you returning to
teaching' and proceeds down the returner path as a UK candidate we end up
exporting their `phone_call_scheduled_at` value incorrectly and the request
fails the API validation.

Instead, we should treat the `UkCallback` step as `skipped?` unless they are
shown the `HaveADegree` step.