Forward request_id to the API
We want to be able to track requests from the app right through to the API. In
order to do this we need to set a correlation id in the request headers going
to the API. Rails has a request id we can tap into, but getting it into the API
client for all requests will be difficult/verbose.

To make life easier this commit exposes the Rails request id to the current
thread. We are using `ActiveSupport::CurrentAttributes` (via
`GetIntoTeachingApiClient::Current`) to ensure the thread globals get cleaned
up after each request.