Work around Sentry::BackgroundWorker dependency on ActiveRecord
The `Sentry::BackgroundWorker` sends events to Sentry async by default, using
`ActiveRecord::Base.connection_pool`. We don't use `ActiveRecord` in the app,
so this was failing silently with an `unrecognized constant ActiveRecord`.
Instead of pulling in `ActiveRecord` just for this (which has its own
difficulties as it tries to connect to a DB by default) this commit overrides
the `async` handler to just send everything synchronously (which is how
`sentry-raven` was working).