Disable VCR for CloudFrontIpFilter tests
I'm not sure why, but these specs don't have VCR casettes so are likely hitting
the actual ip-ranges endpoint during the test run. After installing `WebMock`
these tests began to fail (due to being blocked).

This commit disables VCR and stubs the response to fallback to the cached
ip-ranges versions in the gem.