Swap feedbacks#show for feedbacks#thank_you
As we're not displaying any of the feedback data in the `#show` action view it
makes more sense to have it as a separate `#thank_you` action.

This way, if we did start displaying `Feedback` data on the page we wouldn't
accidently create an avenue for users to retrieve feedback that they did not
directly submit (by plugging in different `id` values into the url path).