Add invisible captcha protection to feedback form
At the moment a bot could easily submit the feedback form; we want to prevent
that so that we only get authentic submissions from candidates.

Add `invisible_captcha` gem, which puts three forms of protection onto the
page:

- A hidden honey pot field (if filled in the submission is rejected).
- A 'spinner' field (if modified the submission is rejected).
- A minimum submission time (too-fast submissions are rejected).

If a submission is rejected we render a HEAD 200 response, which should be
enough to make the bot think it was successful/hopefully it then moves on.