Support multiple basic auth credentials
Extract the basic auth logic from the controller into a lib; updates to support
multiple logins, as we will need this going forward (for restricting access to
the feedback section).

The controller will set a `session[:username]` when auth is successful so that
we can later scope certain actions to particular users.