Do not restrict access to new/create/thank_you actions
The basic auth was being applied to all actions in a production instead of
just the `index` and `export` actions. This went un-noticed in the
rolling/preprod environments due to the site-wide authentication being
expected on all pages anyway.

Only apply basic auth to `index` and `create` actions, _unless_ on a
production-like (but not production) environment.

Don't apply basic auth when in dev/test.