Isolate API client HTTP cache store
The API client library is being updated to support cache invalidation under
certain scenarios (for example, posting a new teaching event needs to clear
the event search cache).

As invalidating specific response caches would be reasonably tricky (the
caching is performed by a separate middleware) the API client instead blows
away its entire cache store (the cache is short-lived and cheap). It is no
longer safe to use `Rails.cache` for the API client and instead we need to
isolate the cache, so we don't invalidate anything apart from cached HTTP
responses.

The TTA service doesn't support any of the endpoints that invalidate cache at
the moment, but we should still isolate the cache to protect against future
chagnes.