Log when a duplicate Candidate is potentially being created
If the API responds with a non-404 error on the match back endpoint then we
are potentially creating a duplicate Candidate record in the CRM (as we have
been unable to verify against an existing Candidate record).

Add logging so that we can easily search for these occurances.

Add logging so that we can easily identify requests for access tokens by
transaction (TTA in this case).