Support alternative completion page copy depending on choices
If a candidate proceeds through as a returner or with an equivalent degree type
we want to show slightly different copy on the completion page.

This was complicated by the fact that we clear the `@store` prior to
redirecting to the completion page; to get around this we persist
`completion_attributes` prior to purging the store and include them as query
parameters in the redirect (none are PII).