Update international telephone validation
When a candidate is offered a callback we need to ensure that their telephone
number will be in a format that the TPUK dialing system supports. On the
client-side this means the number needs to be prefixed by the country dial-in
code (e.g. 44). The API will then sanitize the number further to remove
non-numerical values, white space and a preceeding `+`.