Correct time zone on update action
We were only setting the users time zone on `show` actions, but the
view will be rendered on an `update` action that fails validation
so we need to ensure the correct time zone is applied in that
scenario as well.

Add test to check the time zone offset works for both `show` and
`update` actions.