Update labels for teacher training years
We want to update the labels fo the teacher training years, which come back
from the CRM as '2021', 2022', etc. Instead, if the year is the current year we
want to show the format:

```
<year> - start your training this September
```

We only show the current year up until September. If the year is a future year
we want to use the format:

```
<year> - start your training in September <year>
```