Support custom channel in TTA sign up
The events team are pushing TTA sign ups at on-campus events and want to be
able to attribute a successful sign up to a particular event. The API now
accepts a `channe_id` in the TTA sign up request model, so this PR updates the
TTA root URL to forward on a `channel` attribute, that will be used for the
`channel_id` when signing up. It is persisted as part of the first (identity)
step.

To sign up the URL format is:

```
/?channel=<channel_id>
```