Add BAM tracking pixel to completion page
The TTA service now accepts a custom channel for sign ups so that
they can be attributed to in-person events. In order to analyse the
data we also need the BAM tracking pixel added; this is a lift and
shift from the app and will be replaced with GTM in the future.