Add feature flag to disable analytics/tracking
We are migrating all of our tracking/analytics to a new GTM container.
In order make the switch cleanly we need a way of turning off all our
existing (to be legacy) analytics and turning on our new analytics in
one go.

Add feature flag to control turning off all our existing analytics and
tracking pixels. When the new GTM container is pulled in we will do so
conditionally based on the same feature flag, so that only one or the
other is ever running at once.