Inform GTM of consent changes
As we will be loading all of our JS tracking/analytics snippets in via
GTM going forward we need to be able to set the consent according to
the user's preferences.

When the page loads we send the consent as stored in the cookie
preferernces over to GTM. We also listen to the cookie preference
change event and fire that across if the user updates their
preferences.