Clear non-essential cookies when opting-out
When a user opts-out of a cookie category we want to remove the
cookies associated with activity they have opted-out of. As its not
easy to map the cookies back to the particular categories we will
instead blow away all the non-essential cookies (they will then be
treated as a new user tracking-wise, but we don't expect a lot of
people are doing this in general so its likely not a big deal).