Update to v2 of the API client
The API client v2 uses `snake_case` attributes in the initializers instead
of `camelCase`; this updates all the references.

The initializer now also checks for attributes that don't exist on the
model, raising an exception if any are provided. As some of our steps have
additional attributes to the payload we need to strip these out prior to
initializing the request models and `to_hash` includes `nil` attributes, so
whilst the contract snapshots have been updated they should be logically
equivalent to before still.