Prevent loss of Google Ads tracking parameters
When a user lands on a page via a Google Ads link it will have various query
parameters attached that are used to attribute the visit to an paid ad.

When GA is included on each page (not using GTM) it uses the same 'tracker
object' for each page view. The tracker will use `window.location` on its first
instantiation to attribute the session to a paid ad (via the query parameters).

When GA is included via GTM, however, each time the tag fires it uses a new
tracker object, which looks at `document.location` and ends up attributing
subsequent page views as organic (the query parameters to attribute it to the
ad are only present on the first page view).

To fix this we need to configure the tracker object manually on each page view
to correctly attribute it to the paid ad. We can do this by preserving the
original `document.location` in the `dataLayer` and referencing it in the tag
configuration in GTM.