Handle when invalid channel id is specified
Our on-campus events provide a `channel=` query string parameter (by
way of a QR code/URL). If the channel is invalid we currently show
an unhelpful error message.

Whilst it shouldn't be possible for users to encounter this
scenario, if the QR code author makes a mistake it _could_ lead to
the user running into it.

Handle the error more gracefully by manually clearing the channel and
allowing the user to continue.