Expose sub_channel for BAM tracking pixel
In addition to the `channel` we already record we want to allow BAM to
send a `sub_channel` in the query string so that they can better
attribute sign ups.

Store the `sub_channel` if present in the query string and expose it
via the DOM so that we can retrieve it in GTM and include it in the
magic pixel request.