Update interface to accept setting multiple categories
We want to be able to emit an event when cookies change; currently we use a
`setCategory` method that is called multiple times, so the `CookiePreference`
class doesn't know when all the cookies have finished changing. By changing
this to `setCategories` we encourage callers to set all cookie categories in
one go and can emit an event thereafter.