Send metric when opt-in status changes
When a user updates their opt-in status we want to increment a metric so that
we can track the success rate of opting-in to cookies.

When a user changes their cookie preferences (either manually or via 'allow
all') emit a `tta_client_cookie_consent_total` metric with the respective
labels.