Make change links on review answers page more accessible
Currently the links to change your answer on the review answers page
infer the visually-hidden text from the step name. This is not often a
good experience for the user and results in phrases such as:

"Change identity" instead of "Change your email"

As multiple answers are part of the same step it also means the text is
duplicated which can be confusing.

Instead, the visually-hidden text should relate to the specific answer
being changed.