Skip steps the candidate has seen/submitted already
At the moment the candidate will have to go through the remainder of the wizard
steps when they change an answer from the check answers page.

Instead, we want to jump them straight back to the check answers page if they
have already completed the form.

Pull in updated `dfe_wizard` gem which has this behaviour. Ensure the check
answers step is always displayed to the user by overriding `#seen?`.

Default country to UK during data export instead of within the `uk_or_overseas`
step; if we do it in the step it ends up setting the `country_id` and so the
`overseas_country` step will be skipped if pre-filled when it shouldn't be (as
its not optional).