Use in-memory store in development
If the Rails cache is disabled we were defaulting to a `null_store` for the
cache; we later set the `session_store` to use the `cache_store`, however, so
it was causing the session to immediately expire. Instead, we should use a
`memory_store` in development (caching is disabled by setting `perform_caching`
to false).

I'm not sure how this ever worked :/