Include StageInterestedTeaching step in returning teacher journey
Currently, if a candidate is a returning teacher we don't ask them what their
preferred stage of teaching is (primary or secondary). Instead we disqualify
some candidates based on the subject they want to teach (if its not one of a
small subset of secondary subjects).

Going forward we want to allow all secondary subjects for returning teachers
and only disqualify them from the journey if they want to teach primary. The
first step for this is to include the `StageInterestedTeaching` step in the
returners journey.

A few other steps that used the state of the `StageInterestedTeachinga` step to
determine if they should be skipped have been updated to ensure other journies
remain consistent.

The contract tests are updated to include the `preferredEducationPhaseId` for
returning teacher submissions and feature tests updated to include the
additional step for returners.