Move education stage exit step to be after StageInterestedTeaching
We previously had an exit step `SubjectNotFound` which the candidate would hit
if they chose 'Other' as the subject they wanted to teach. We now ask them
their preferred education phase (primary or secondary) and we can disqualify
them based on this as we only accept secondary subjects.

If the candidate chooses 'primary' take them to the `PrimaryReturner` exit
step.