Add dfe-analytics gem
Add the `dfe-analytics` gem and enable it when running locally. Capture all web
requests and model updates as events (we only have a single `Feedback` model we
are capturing).

Minor tweak to the `request_id` specs to allow `request.uuid` to be called
multiple times; it appears the `dfe-analytics` gem also requests the `uuid` so
we need to allow the mock to be called more than once.

The JSON secret for BigQuery has been added to the Azure keyvault.

Enabled in all environments apart from production to begin with.