Update copy on teacher training years options
We want to update the copy around the first available year in the set of
options depeneding on what the current date is. This is in an effort to clear
up the reference to "this September" when the date is after the cutoff (7th
September) and "this September" is actually "next September" until January 1st
when it becomes "this September" again!

The logic is as follows, assuming the current year is 2022:

1st January 2022 to 23rd June 2022:

```
Not sure
2022 - start your training this September
2023
2024
```

between 24th June 2022 and 6th September 2022:

```
Not sure
2022 - start your training this September
2023
2024
2025
```

between 7th September 2022 and 31st December 2022:

```
Not sure
2023 - start your training next September
2024
2025
```

between 1st January 2023 and 23rd June 2023:

```
Not sure
2023 - start your training this September
2024
2025
```

... etc ...