Fix flakey integration test
The integration test for an existing candidate is being flakey; I'm not certain
why yet but I think it may be to do with Capybara not resetting the session
between tests consistently.

Update the existing candidate test to simplify it; we really only care that the
candidate gets a verification email and can proceed to the next step. The new
candidate test will still ensure the submission to the API/CRM is working.