Update dfe-analytics gem
Update to the latest version of `dfe-analytics`. There is an issue whereby
the latest gem calls `eager_load!` which causes Rails to evaluate the
`inclusion` validation block in `WhatDegreeClass` and make an external
request (which fails due to being unauthorized in the test environment). To
prevent this we supress the dfe-analytics initialisation process.