Remove privacy policy step from sign up journey
We have a link and text to call out the privacy notice on the first page of the
sign up journey; there is not a requirement to have the user explicitly accept
the privacy policy and so we can remove that step from the journey.

We still need to send the API/CRM the ID of the privacy notice we surfaced to
the user during sign up, however, so we now capture this on the first step.