Gracefully handle when there are no available callback slots
Currently the API will provide a fallback list of available callback slots
in the event the CRM returns none. Going forward the API will also return 0
if the CRM has none available.

We need to account for this in the front-end by asking the candidates to
contact us instead of providing a mechanism for them to book a callback.

Update copy on steps that ask for a callback; on the overseas journey its
actually two steps as we ask for their phone number/time zone before showing
a list of callback slots (in their selected time zone). If the candidate is
unable to book a callback we will not ask for their telephone number/time
zone.