Skip SubjectInterestedTeaching step for students
We no longer want to ask students for the subject they are
interested in teaching. This step was actually removed
when we skipped the `StageInterestedTeaching` step as they
were inter-dependent already (you had to select Secondary
in order to choose a subject you want to teach), but this
change makes it explicit.